% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{US Cities}
\format{
A tibble with 10000 rows and 11 variables:
}
\source{
\url{https://www.latlong.net/category/cities-236-15.html}
}
\usage{
cities
}
\description{
A dataset containing the location of 10,000 US cities
}
\keyword{datasets}
