% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddist.R
\name{ddist}
\alias{ddist}
\title{Add a DSB style to your graph}
\usage{
ddist(data = NULL, id = NULL, latitude = NULL, longitude = NULL, crs = 4326)
}
\arguments{
\item{...}{x}
}
\value{
... y
}
\description{
This function etc.
}
\author{
Jeppe Vierø
}
