% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddist_sf.R
\name{ddist_sf}
\alias{ddist_sf}
\title{Create dyads of points i and j and calculate distances}
\usage{
ddist_sf(
  data = NULL,
  id = NULL,
  crs_transform = FALSE,
  new_crs = NULL,
  diagonal = TRUE
)
}
\arguments{
\item{data}{an sf data.frame or tibble.}

\item{id}{a numeric variable uniquely idenfiying points.}

\item{crs_transform}{a logical value indicating whether to transform the Coordinate Reference System (CRS). Defaults to FALSE}

\item{new_crs}{a valid EPSG for a new CRS.}

\item{diagonal}{a logical value indicating whether to keep the diagonal component in the distance matrix with dyads (i,i). Defaults to TRUE}
}
\value{
... y
}
\description{
This function calculates the geodesic distance between any dyads (pairs of points) and stores the result in a long tibble, a opposed to a wide matrix.
}
\author{
Jeppe Vierø
}
