% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddist_xy_sf.R
\name{ddist_xy_sf}
\alias{ddist_xy_sf}
\title{Create dyads of points i and j and calculate distances}
\usage{
ddist_xy_sf(
  x = NULL,
  y = NULL,
  id_x = NULL,
  id_y = NULL,
  crs_transform = FALSE,
  new_crs = NULL
)
}
\arguments{
\item{x}{an object of class \code{data.frame} or \code{tibble}.}

\item{y}{an object of class \code{data.frame} or \code{tibble}.}

\item{id_x}{a variable uniquely idenfiying geospatial points in data \code{x}. Can be of type numeric, integer, character, or factor}

\item{id_y}{a variable uniquely idenfiying geospatial points in data \code{y}. Can be of type numeric, integer, character, or factor}

\item{crs_transform}{a logical value indicating whether to transform the CRS. Defaults to FALSE}

\item{new_crs}{a valid EPSG for a new CRS. See \code{rgdal::make_EPSG()} or \url{https://epsg.org/home.html}}
}
\value{
a long \link[tibble]{tibble} with dyads and dyadic distances incl. a distance unit indicator
}
\description{
This function calculates the geodesic distance between any dyads (pairs of points) and stores the result in a long tibble, a opposed to a wide matrix.
}
\author{
Jeppe Vierø
}
