% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.R
\name{pdist}
\alias{pdist}
\title{Create dyads of points i and j and calculate distances}
\usage{
pdist(
  x = NULL,
  y = NULL,
  id = NULL,
  longitude = "longitude",
  latitude = "latitude",
  crs = 4326,
  crs_transform = FALSE,
  new_crs = NULL
)
}
\arguments{
\item{x}{an object of class \code{data.frame} or \code{tibble}.}

\item{y}{an object of class \code{data.frame} or \code{tibble}.}

\item{id}{a variable uniquely idenfiying geospatial points. Can be of type numeric, integer, character, or factor}

\item{longitude}{name of the numeric longitude variable. Defaults to c("longitude", "longitude")}

\item{latitude}{name of the numeric latitude variable. Defaults to c("latitude", "latitude")}

\item{crs}{a valid EPSG for a valid Coordinate Reference System (CRS) for your coordinates. Defaults to c(4326, 4326).}

\item{crs_transform}{a logical value indicating whether to transform the CRS'. Defaults to FALSE}

\item{new_crs}{a valid EPSG for a new CRS. See \code{rgdal::make_EPSG()} or \url{https://epsg.org/home.html}}
}
\value{
a long \link[tibble]{tibble} with dyads and dyadic distances incl. a distance unit indicator
}
\description{
This function calculates the geodesic distance between any dyads (pairs of points) and stores the result in a long tibble, a opposed to a wide matrix.
}
\examples{
library(tibble)
df <- tibble::tribble(
~city_name, ~idvar, ~latitude, ~longitude,
"copenhagen", 5, 55.68, 12.58,
"stockholm", 2, 59.33, 18.07,
"oslo", 51, 59.91, 10.75
)
pdist(data = df, id = "idvar")
}
\author{
Jeppe Vierø
}
