% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa}
\alias{usa}
\title{USA}
\format{
An \code{sf} tibble with 1 row and 2 variables:
}
\source{
\url{https://www.naturalearthdata.com/}
}
\usage{
usa
}
\description{
A spatial dataset containing
}
\keyword{datasets}
