% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddist.R
\name{ddist}
\alias{ddist}
\title{Create dyads of points i and j and calculate distances}
\usage{
ddist(data = NULL, id = NULL, crs = 4326)
}
\arguments{
\item{data}{a data.frame or tibble}
}
\value{
... x
}
\description{
This function calculates the geodesic distance between any dyads (pairs of points) and stores the result in a long tibble, a opposed to a wide matrix
}
\author{
Jeppe Vierø
}
