% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddist.R
\name{ddist}
\alias{ddist}
\title{Create dyads of points i and j and calculate distances}
\usage{
ddist(
  data = NULL,
  id = NULL,
  longitude = "longitude",
  latitude = "latitude",
  crs = 4326,
  crs_transform = FALSE,
  new_crs = NULL,
  diagonal = TRUE,
  duplicates = TRUE
)
}
\arguments{
\item{data}{a data.frame or tibble.}

\item{id}{a variable uniquely idenfiying geospatial points. Can be of type numeric, integer, character, or factor}

\item{longitude}{name of the numeric longitude variable. Defaults to "longitude"}

\item{latitude}{name of the numeric latitude variable. Defaults to "latitude"}

\item{crs}{a valid EPSG for a valid Coordinate Reference System (CRS) for your coordinates. Defaults to 4326.}

\item{crs_transform}{a logical value indicating whether to transform the CRS. Defaults to FALSE}

\item{new_crs}{a valid EPSG for a new CRS. See \url{https://epsg.org/home.html}}

\item{diagonal}{a logical value. Keep the diagonal component in the distance matrix with dyads (i,i) and distance zero? Defaults to TRUE}

\item{duplicates}{a logical value. Keep "identical" dyads (i,j) and (j,i)? Defaults to TRUE. If set to FALSE, only one observation per dyad is kept. Note that this uses \link[rowwise]{dplyr} and makes the \link[ddist]{dyadicdist} function considerably more time-consuming}
}
\value{
a long \link[tibble]{tibble} with dyads and dyadic distances
}
\description{
This function calculates the geodesic distance between any dyads (pairs of points) and stores the result in a long tibble, a opposed to a wide matrix.
}
\author{
Jeppe Vierø
}
