% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddist_sf.R
\name{ddist_sf}
\alias{ddist_sf}
\title{Calculate dyadic distances between points \code{i} and \code{j} in a spatial dataset}
\usage{
ddist_sf(
  data = NULL,
  id = NULL,
  crs_transform = FALSE,
  new_crs = NULL,
  diagonal = TRUE,
  duplicates = TRUE
)
}
\arguments{
\item{data}{an object of class \code{sf} (\verb{"sf" "data.frame"} or \verb{"sf" "tbl_df" "tbl" "data.frame"})}

\item{id}{a variable uniquely idenfiying geospatial points. Can be of type numeric, integer, character, or factor}

\item{crs_transform}{a logical value indicating whether to transform the CRS. Defaults to FALSE}

\item{new_crs}{a valid EPSG for a new CRS. See \code{rgdal::make_EPSG()} or \url{https://epsg.org/home.html}}

\item{diagonal}{a logical value. Keep the diagonal component in the distance matrix with dyads (i,i) and distance zero? Defaults to TRUE}

\item{duplicates}{a logical value. Keep "identical" dyads (i,j) and (j,i)? Defaults to TRUE. If set to FALSE, only one observation per dyad is kept. Note that this uses \link[dplyr]{rowwise} and makes the \link[dyadicdist]{ddist} function considerably more time-consuming}
}
\value{
a long \link[tibble]{tibble} with dyads and dyadic distances incl. a distance unit indicator
}
\description{
This function calculates the geodesic distance between any dyads (pairs of points \code{i} and \code{j}) in a spatial dataset and stores the result in a long tibble
}
\examples{
library(tibble)
df <- tibble::tribble(
~city_name, ~idvar, ~latitude, ~longitude,
"copenhagen", 5, 55.68, 12.58,
"stockholm", 2, 59.33, 18.07,
"oslo", 51, 59.91, 10.75
) \%>\% sf::st_as_sf(coords = c("longitude", "latitude"), crs = 4326)
ddist_sf(data = df, id = "idvar")
}
\author{
Jeppe Vierø
}
