% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddist_xy.R
\name{ddist_xy}
\alias{ddist_xy}
\title{Calculate dyadic distances between points \code{i} and \code{j} from separate datasets}
\usage{
ddist_xy(
  x = NULL,
  y = NULL,
  ids = NULL,
  coords_x = c("longitude", "latitude"),
  coords_y = c("longitude", "latitude"),
  crs = c(4326, 4326),
  crs_transform = FALSE,
  new_crs = NULL
)
}
\arguments{
\item{x}{an object of class \code{data.frame} or \code{tibble}.}

\item{y}{an object of class \code{data.frame} or \code{tibble}.}

\item{crs}{valid EPSG's for your coordinates in data \code{x} and \code{y}. Defaults to c(4326, 4326).}

\item{crs_transform}{a logical value indicating whether to transform the CRS. Defaults to FALSE}

\item{new_crs}{a valid EPSG for a new CRS. See \code{rgdal::make_EPSG()} or \url{https://epsg.org/home.html}}

\item{id}{variables uniquely identifying geospatial points in data \code{x} and \code{y}. Can be of type numeric, integer, character, or factor}

\item{longitude}{names of the numeric longitude variables in data \code{x} and \code{y}. Defaults to c("longitude", "longitude")}

\item{latitude}{names of the numeric latitude variables in data \code{x} and \code{y}. Defaults to c("latitude", "latitude")}
}
\value{
a long \link[tibble]{tibble} with dyads and dyadic distances incl. a distance unit indicator
}
\description{
This function calculates the geodesic distance between any dyads (pairs of points \code{i} and \code{j}) from datasets \code{x} and \code{y} and stores the result in a long tibble
}
\author{
Jeppe Vierø
}
