% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddist_xy.R
\name{ddist_xy}
\alias{ddist_xy}
\title{Create dyads of points i and j and calculate distances}
\usage{
ddist_xy(
  x = NULL,
  y = NULL,
  x_id = NULL,
  y_id = NULL,
  x_longitude = longitude,
  x_latitude = latitude,
  y_longitude = longitude,
  y_latitude = latitude,
  x_crs = 4326,
  y_crs = 4326,
  crs_transform = FALSE,
  new_crs = NULL
)
}
\arguments{
\item{x}{an object of class \code{data.frame} or \code{tibble}.}

\item{y}{an object of class \code{data.frame} or \code{tibble}.}

\item{x_id}{a variable uniquely idenfiying geospatial points in data \code{x}. Can be of type numeric, integer, character, or factor}

\item{y_id}{a variable uniquely idenfiying geospatial points in data \code{y}. Can be of type numeric, integer, character, or factor}

\item{x_longitude}{name of the numeric longitude variable in data \code{x}. Defaults to "longitude"}

\item{x_latitude}{name of the numeric latitude variable in data \code{x}. Defaults to "latitude"}

\item{y_longitude}{name of the numeric longitude variable in data \code{y}. Defaults to "longitude"}

\item{y_latitude}{name of the numeric latitude variable in data \code{y}. Defaults to "latitude"}

\item{x_crs}{a valid EPSG for a valid Coordinate Reference System (CRS) for your coordinates in data \code{x}. Defaults to 4326.}

\item{y_crs}{a valid EPSG for a valid Coordinate Reference System (CRS) for your coordinates in data \code{y}. Defaults to 4326.}

\item{crs_transform}{a logical value indicating whether to transform the CRS. Defaults to FALSE}

\item{new_crs}{a valid EPSG for a new CRS. See \code{rgdal::make_EPSG()} or \url{https://epsg.org/home.html}}
}
\value{
a long \link[tibble]{tibble} with dyads and dyadic distances incl. a distance unit indicator
}
\description{
This function calculates the geodesic distance between any dyads (pairs of points) and stores the result in a long tibble, a opposed to a wide matrix.
}
\author{
Jeppe Vierø
}
